+++ R/group_grid.R
@@ -15,7 +15,6 @@ group_grid <- function(x, i = NULL, j = NULL, ...) {
 
 group_grid_col <- function(x, j, ...) {
   m <- meta(x)
-  # columns
   header <- empty_cells(j)
   cw <- meta(x, "col_widths")
   cw <- sapply(header, function(k) sum(cw[k]) + length(cw[k]) - 1)
@@ -24,10 +23,11 @@ group_grid_col <- function(x, j, ...) {
   out <- strsplit(out, split = "\\n")[[1]]
   out <- out[out != "\\n"]
   out <- out[!out %in% c("\\n", "")]
+  out <- out[2]
   x <- strsplit(x, split = "\\n")[[1]]
   x <- x[!x %in% c("\\n", "")]
-  out <- out[1:(length(out) - 1)]
-  out <- paste(c(out, x), collapse = "\n")
+  x <- c(x[1], out, x)
+  out <- paste(x, collapse = "\n")
   attr(out, "tinytable_meta") <- m
   class(out) <- class(x)
   return(out)