+++ R/tt.R
@@ -1,23 +1,30 @@
 #' Draw a Tiny Table
 #'
 #' @description
-#' The `tt` function renders a table in different formats (HTML, Markdown, or LaTeX) with various styling options. The table can be customized with additional functions:
+#' The `tt` function renders a table in different formats with various styling options: HTML, Markdown, LaTeX, Word, PDF, PNG, or Typst. The table can be customized with additional functions:
 #'
 #' * `style_tt()` to style fonts, colors, alignment, etc.
 #' * `format_tt()` to format numbers, dates, strings, etc.
+#' * `group_tt()` for row or column group labels.
 #' * `save_tt()` to save the table to a file or return the table as a string.
+#' * `print()` to print to a specific format, ex: `print(x, "latex")`
 #'
 #' `tinytable` attempts to determine the appropriate way to print the table based on interactive use, RStudio availability, and output format in RMarkdown or Quarto documents. Users can call `print(x, output="markdown")` to print the table in a specific format. Alternatively, they can set a global option: `options("tinytable_print_output"="markdown")`
 #'
 #' @param x A data frame or data table to be rendered as a table.
-#' @param digits Number of significant digits to keep for numeric variables. When `digits` is an integer, `tt()` calls `format_tt(x, digits = digits)` before proceeding to draw the table. Users who need more control can proceed in two steps: (1) format the data with `format_tt()` or other functions, and (2) pass the formatted data to `tt()` for drawing. See `?format_tt` for more details on formating options (ex: decimal, scientific notation, dates, boolean variables, etc.).
+#' @param digits Number of significant digits to keep for numeric variables. When `digits` is an integer, `tt()` calls `format_tt(x, digits = digits)` before proceeding to draw the table. Users who need more control can proceed in two steps: (1) format the data with `format_tt()` or other functions, and (2) pass the formatted data to `tt()` for drawing. See `?format_tt` for more details on formatting options (ex: decimal, scientific notation, dates, boolean variables, etc.).
 #' @param caption A string that will be used as the caption of the table.
 #' @param width A numeric value between 0 and 1 indicating the proportion of the line width that the table should cover.
 #' @param theme The theme to apply to the table: "default", "striped", "bootstrap", "void", or "grid".
-#' @param notes A single string or a (named) list of strings to append at the bottom of the table.
-#' 
+#' @param notes Notes to append to the bottom of the table. This argument accepts several different inputs:
+#' * Single string insert a single note: `"blah blah"`
+#' * Multiple strings insert multiple notes sequentially: `list("Hello world", "Foo bar")`
+#' * A named list inserts a list with the name as superscript: `list("a" = list("Hello World"))`
+#' * A named list with positions inserts markers as superscripts inside table cells: `list("a" = list(i = 0:1, j = 2, text = "Hello World"))`
 #' @param placement A string to control the position of tables in LaTeX. Will be inserted in square brackets like: `\\begin{table}[H]`
 #' @return An object of class `tt` representing the table.
+#' 
+#' The table object has an attribute which holds information about the structure of the table. This metadata can be accessed with `attr(x,"tinytable_meta")`. In general, modifying the content of this attribute is not recommended, but it can be useful to some developers, such as those who want to force print to a specific output format without calling `print()`.
 #' @template latex_preamble
 #' 
 #' @examples
@@ -30,7 +37,13 @@
 #'    theme = "striped",
 #'    width = 0.5,
 #'    caption = "Data about cars.")
+#' 
+#' tt(x, notes = "Hello World!")
 #'
+#' fn <- list(i = 0:1, j = 2, text = "Hello World!")
+#' tab <- tt(x, notes = list("*" = fn))
+#' print(tab, "latex")
+#' 
 #' @export
 tt <- function(x,
                digits = NULL,
@@ -51,10 +64,7 @@ tt <- function(x,
 
   
   # notes can be a single string or a (named) list of strings
-  if (is.character(notes) && length(notes)) {
-    notes <- list(notes)
-  }
-  assert_list(notes, null.ok = TRUE)
+  sanity_notes(notes)
 
   # before style_tt() call for align
   out <- x 
@@ -67,22 +77,13 @@ tt <- function(x,
   out <- meta(out, "nhead", if (is.null(colnames(x))) 0 else 1)
   out <- meta(out, "nrows", nrow(x))
   out <- meta(out, "ncols", ncol(x))
+  out <- meta(out, "notes", notes)
   out <- meta(out, "caption", caption)
   class(out) <- c("tinytable", class(out))
 
   # build table
+  # tt_tabularray wil be substituted in build_tt by the appropriate on based on output
   cal <- call("tt_tabularray", x = out, caption = caption, theme = theme, width = width, notes = notes, placement = placement)
-  # if (output == "latex") {
-  #
-  # } else if (output == "html"){
-  #   cal <- call("tt_bootstrap", x = out, caption = caption, theme = theme, width = width, notes = notes)
-  #
-  # } else if (output == "markdown") {
-  #   cal <- call("tt_grid", x = out, caption = caption)
-  #
-  # } else {
-  #   stop("here be dragons")
-  # }
 
   out <- meta(out, "lazy_tt", cal)